+++ tests/scanner_util_test.h
@@ -1,0 +1,12 @@
+#ifndef scanner_util_test_h
+#define scanner_util_test_h
+
+#include <stdio.h>
+#include "CuTest.h"
+#include "scanner_util.h"
+
+
+CuSuite *scannerUtilGetSuite();
+
+
+#endif